Fix lint error: rename feature_toggle to featuretoggle
- Rename package from feature_toggle to featuretoggle to fix stylecheck warning
- Update imports and function calls in config.go
- Package names should not use underscores per Go conventions
